# swagger.yaml
openapi: 3.0.0
paths:
...
  /languages:
    get:
      operationId: getLanguages
      parameters:
        - name: pageSize
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 20
            type: number
        - name: pageIndex
          required: false
          in: query
          description: Page index (starts from 0)
          schema:
            example: 0
            type: number
        - name: order
          required: false
          in: query
          description: Sort order
          schema:
            example: asc
            type: string
            enum:
              - asc
              - desc
        - name: orderBy
          required: false
          in: query
          description: Field to order by
          schema:
            example: createdAt
            type: string
            enum:
              - createdAt
              - name
              - updatedAt
        - name: keyword
          required: false
          in: query
          description: Search keyword
          schema:
            example: admin
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retrieve a list of languages with pagination.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/LanguageResponseDto"
                    required:
                      - data
      summary: Get a list of languages
      tags: &a2
        - Languages
  /languages/{id}:
    get:
      operationId: getLanguageById
      parameters:
        - name: id
          required: true
          in: path
          description: Language ID (ISO 639-1)
          schema:
            minLength: 2
            maxLength: 2
            example: en
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retrieve a specific language by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguageResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Get a language by ID
      tags: *a2
    put:
      operationId: updateLanguage
      parameters:
        - name: id
          required: true
          in: path
          description: Language ID (ISO 639-1)
          schema:
            minLength: 2
            maxLength: 2
            example: en
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageUpdateRequestDto"
      responses:
        "200":
          description: Update an existing language.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguageUpdateResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Update a language
      tags: *a2
    delete:
      operationId: deleteLanguage
      parameters:
        - name: id
          required: true
          in: path
          description: Language ID (ISO 639-1)
          schema:
            minLength: 2
            maxLength: 2
            example: en
            type: string
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete a specific language by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguageDeleteResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Delete a language
      tags: *a2
  /languages/create:
    post:
      operationId: createLanguage
      parameters:
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageCreateRequestDto"
      responses:
        "200":
          description: Create a new language.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguageCreateResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Create a new language
      tags: *a2
...
info:
  title: Ecom API
  description: The Ecom API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    PaginationResponseDto:
      type: object
      properties:
        totalPages:
          type: number
          description: Total number of pages
          example: 5
        totalItems:
          type: number
          description: Total number of items
          example: 100
        pageSize:
          type: number
          description: Number of items per page
          example: 20
        pageIndex:
          type: number
          description: Current page index (starts from 0)
          example: 0
      required:
        - totalPages
        - totalItems
        - pageSize
        - pageIndex
    PageDto:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/PaginationResponseDto"
      required:
        - pagination
    LanguageResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
        name:
          type: string
          description: Language name
          example: English
      required:
        - id
        - name
    LanguageCreateRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Language name
          example: English
          minLength: 2
          maxLength: 500
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
          minLength: 2
          maxLength: 2
      required:
        - name
        - id
    LanguageCreateResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
        name:
          type: string
          description: Language name
          example: English
      required:
        - id
        - name
    LanguageUpdateRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Language name
          example: Vietnamese
          minLength: 2
          maxLength: 500
      required:
        - name
    LanguageUpdateResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
        name:
          type: string
          description: Language name
          example: English
      required:
        - id
        - name
    LanguageDeleteResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
        name:
          type: string
          description: Language name
          example: English
      required:
        - id
        - name
...